Fix warnings on 64-bit platforms about casts from pointers to ints. Two of these were real bugs.

+++ Objects/unicodeobject.c
@@ -6465,7 +6465,8 @@ PyObject *PyUnicode_Format(PyObject *format,
 			     "unsupported format character '%c' (0x%x) "
 			     "at index %i",
 			     (31<=c && c<=126) ? c : '?', 
-                             c, fmt -1 - PyUnicode_AS_UNICODE(uformat));
+                             c,
+			     (int)(fmt -1 - PyUnicode_AS_UNICODE(uformat)));
 		goto onError;
 	    }
 	    if (sign) {
